refactor: Controller 및 명령어 수행 로직 변경
